requirements.txt
fastapi
uvicorn
websockets
dashscope==1.19.0
jinja2
python-multipart
scipy
numpy
python-dotenv
aiofiles
psutil

